feat: ForumOceanWeb add method: getRecommendedClub
fix: 會員數LOng->Int

rename

+++ backend2/src/main/java/com/cmoney/backend2/forumocean/service/api/schemas/v2/RecommendedClub.kt
@@ -35,5 +35,5 @@ data class RecommendedClub(
 
     /** 會員數 */
     @SerializedName("memberCount")
-    val memberCount: Long
+    val memberCount: Int
 )
